% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AND.R
\name{AND}
\alias{AND}
\title{Determine if Two Conditions are Both True}
\usage{
AND(condition_1, condition_2)
}
\arguments{
\item{condition_1}{First condition that you would like evaluated.}

\item{condition_2}{Second condition that you would like evaluated.}
}
\value{
Returns true if both conditions are true. Otherwise returns false.
}
\description{
Evaluates both conditions to determines if both of them are true.
}
\examples{
AND(2 == 2, 4 == 4)
AND(2 == 2, 4 != 4)
AND(2 != 2, 4 != 4)

}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, January 2021
}
\keyword{and}
