% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AVERAGE.R
\name{AVERAGE}
\alias{AVERAGE}
\title{Calculates the mean value of the vector}
\usage{
AVERAGE(vector)
}
\arguments{
\item{vector}{Vector that you would the mean calculated from.}
}
\value{
Returns the mean of the vector as a numeric or an integer.
}
\description{
Concatenates a vector of integers or numerics.
}
\examples{
AVERAGE(c(3, 1, 3))

}
\seealso{
\code{\link{mean}}
}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, December 2020
}
\keyword{mean}
