% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AVERAGEIF.R
\name{AVERAGEIF}
\alias{AVERAGEIF}
\title{Averages Vector Components Based On Condition}
\usage{
AVERAGEIF(avg_vector, condition_vector, condition)
}
\arguments{
\item{avg_vector}{Vector that you would like to average.}

\item{condition_vector}{Vector that you would like condition specified for.}

\item{condition}{Value that will satisfy certain components in condition vector.}
}
\value{
Returns average in the average vector based on condition specified for condition vector.
}
\description{
Takes two vectors, one being the vector you want to average, and the other
being the vector you want create a condition for. Then, specify the condition
that should be used to satisfy the condition vector and it will take the average of the 
average vector accordingly.
}
\examples{
AVERAGEIF(iris$Sepal.Length, iris$Species, "virginica")
AVERAGEIF(c(1, 2, 3), c(3, 1, 3), 3)

}
\seealso{
\code{\link{mean}}
}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, July 2021
}
\keyword{average}
\keyword{conditional}
