% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONCATENATE.R
\name{CONCATENATE}
\alias{CONCATENATE}
\title{Concatenate Vector Components}
\usage{
CONCATENATE(vector)
}
\arguments{
\item{vector}{Vector that you would like concatenated.}
}
\value{
Returns the concatenated vector in the form of a single string.
}
\description{
Concatenates a vector of integers, strings, numerics, or factors.
}
\examples{
CONCATENATE(c("hi", "hey", "yes"))
CONCATENATE(c(3, 1, 3))

}
\seealso{
\code{\link{paste0}}
}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, September 2020
}
\keyword{paste}
