% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORREL.R
\name{CORREL}
\alias{CORREL}
\title{Calculate Correlation Between Two Vectors}
\usage{
CORREL(vec1, vec2)
}
\arguments{
\item{vec1}{Vector of integers or numerics.}

\item{vec2}{Vector of integers or numerics.}
}
\value{
Returns the correlation coefficient between the two vectors.
}
\description{
Given two vectors, return the correlation coefficient between them.
}
\examples{
CORREL(c(1, 2, 4, 2), c(2, 5, 7, 2))
CORREL(c(1, 2, 4, 2, 8, 10, 12), c(2, 5, 7, 2, 20, 24, 12))

}
\seealso{
\code{\link{cor}}
}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, July 2021
}
\keyword{cor}
\keyword{correlation,}
