% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COUNTBLANK.R
\name{COUNTBLANK}
\alias{COUNTBLANK}
\title{Count Number of Blanks in Vector}
\usage{
COUNTBLANK(count_vector)
}
\arguments{
\item{count_vector}{Vector that you would like counted for blanks.}
}
\value{
Returns the count of blanks in vector as an integer.
}
\description{
Given a vector, counts the number of blanks ("") contained in the vector.
}
\examples{
COUNTBLANK(c("", "", "hey", "home"))
COUNTBLANK(c("", 0, 1, 2))

}
\seealso{
\code{\link{length}}
}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, July 2021
}
\keyword{blank}
\keyword{conditional}
\keyword{count}
