% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAYS.R
\name{DAYS}
\alias{DAYS}
\title{Difference in Days Between Two Dates}
\usage{
DAYS(end_date, start_date)
}
\arguments{
\item{end_date}{End date in format "yyyy-mm-dd" as string.}

\item{start_date}{Start date in format "yyyy-mm-dd" as string.}
}
\value{
Returns the number of days between the start and end date.
}
\description{
Given a start and end date, calculate the difference in days between the two.
}
\examples{
DAYS("2016-01-02", "2016-01-01")
DAYS("2017-01-01", "2018-05-28")

}
\seealso{
\code{\link{as.Date}}
}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, January 2021
}
\keyword{Date}
\keyword{Days}
\keyword{Difference}
