% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXACT.R
\name{EXACT}
\alias{EXACT}
\title{Test if Two Strings or Vectors of Strings are Equal}
\usage{
EXACT(text1, text2)
}
\arguments{
\item{text1}{String or vector that you would like to compare to text2.}

\item{text2}{String or vector that you would like to compare to text1.}
}
\value{
Returns boolean value(s) that tell if exact match or not.
}
\description{
Takes either two strings or two vectors that can be converted to strings
and compares to see if they are an exact match.
}
\examples{
EXACT("hi", "hey")
EXACT(c("hey", "hi", "holly"), c("hey", "HI", "holly"))

}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, July 2021
}
\keyword{equal}
\keyword{exact}
\keyword{same}
