% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIND.R
\name{FIND}
\alias{FIND}
\title{Find Character Number in String}
\usage{
FIND(char_find, x)
}
\arguments{
\item{char_find}{Specific character you would like to find in string/vector.}

\item{x}{Character vector or string you would like to slice.}
}
\value{
Returns the first character number of a specified character in a string or character vector.
}
\description{
Returns the character number of a specified character in a string/vector.
}
\examples{
FIND("l", "hello")
FIND("l", c("hello", "hey", "hi", "yellow"))

}
\seealso{
\code{\link{strsplit}}
}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, January 2021
}
\keyword{find}
