% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FV.R
\name{FV}
\alias{FV}
\title{Calculate Future Value of Cash Flows}
\usage{
FV(rate, nper, pmt, pv)
}
\arguments{
\item{rate}{Interest rate you would like used when compounding.}

\item{nper}{Number of periods to use when compounding the present value.}

\item{pmt}{Vector of future cash flows (should match nper) that you would like compounded. Starts at t = 1.}

\item{pv}{Present amount you wish to compound.}
}
\value{
Returns the future value of cash flows given.
}
\description{
Given an interest rate, the number of payment periods, the periodic payments, and the present value, calculate the future value.
}
\examples{
FV(0.05, 5, c(50, 50, 50, 50, 50), 100)
FV(0.01, 3, c(50, 100, 150), 100)
FV(0.03, 10, 0, 5000)

}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, March 2021
}
\keyword{discounting}
\keyword{future}
\keyword{value}
