% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEFT.R
\name{LEFT}
\alias{LEFT}
\title{Return Specified Number of Characters from Left in String/Vector}
\usage{
LEFT(x, length)
}
\arguments{
\item{x}{Character vector or string you would like to slice.}

\item{length}{Number of characters from left you would like returned.}
}
\value{
Returns the number of characters from left in a string/character vector
}
\description{
Returns the number of characters from left in a string/character vector
}
\examples{
LEFT("hello world!", 5)
LEFT(c("hi", "hey", "hello"), 2)

}
\seealso{
\code{\link{substr}}
}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, January 2021
}
\keyword{substr}
