% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEN.R
\name{LEN}
\alias{LEN}
\title{Count Characters in String or Vector}
\usage{
LEN(text)
}
\arguments{
\item{text}{String or vector that you would like its characters counted.}
}
\value{
Returns the number of characters in string or vector.
}
\description{
Given a string or vector, count the number of characters for each string.
}
\examples{
LEN("hi")
LEN(c("hey", "hi", "hey"))
LEN(c(1, "hi", "hey", 2))

}
\seealso{
\code{\link{nchar}}
}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, February 2021
}
\keyword{count}
\keyword{len}
\keyword{nchar}
