% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOWER.R
\name{LOWER}
\alias{LOWER}
\title{Convert Text to Lowercase}
\usage{
LOWER(text)
}
\arguments{
\item{text}{String or vector that you would like converted to lowercase text.}
}
\value{
Returns the string or vector in lowercase text.
}
\description{
Given a string or vector, convert to lowercase text.
}
\examples{
LOWER("HELLO")
LOWER(c("hey", "HI", "HEY"))
LOWER(c(1, "HI", "HEY", "hi"))

}
\seealso{
\code{\link{tolower}}
}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, January 2021
}
\keyword{lower}
\keyword{tolower}
