% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOW.R
\name{NOW}
\alias{NOW}
\title{Get Today's Date & Time}
\usage{
NOW()
}
\value{
Returns today's date and time as a date-time object.
}
\description{
Returns today's date and time.
}
\examples{
NOW()

}
\seealso{
\code{\link{Sys.time}}
}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, February 2021
}
\keyword{now}
\keyword{time}
\keyword{today}
