% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OR.R
\name{OR}
\alias{OR}
\title{Determine if One of Two Condistions are True}
\usage{
OR(condition_1, condition_2)
}
\arguments{
\item{condition_1}{First condition that you would like evaluated.}

\item{condition_2}{Second condition that you would like evaluated.}
}
\value{
Returns true if at least one of the conditions are true. Otherwise returns false.
}
\description{
Evaluates both conditions to determines if one of them is true.
}
\examples{
OR(2 == 2, 4 == 4)
OR(2 == 2, 4 != 4)
OR(2 != 2, 4 != 4)

}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, January 2021
}
\keyword{or}
