% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PV.R
\name{PV}
\alias{PV}
\title{Calculate Present Value of Cash Flows}
\usage{
PV(rate, nper, pmt, fv)
}
\arguments{
\item{rate}{Interest rate you would like used when discounting.}

\item{nper}{Number of periods to use when discounting the future value.}

\item{pmt}{Vector of payments (should match nper).}

\item{fv}{Future amount you wish to discount.}
}
\value{
Returns the present value of cash flows given.
}
\description{
Given an interest rate, the number of payment periods, the periodic payments, and the future value, calculate the present value.
}
\examples{
PV(0.05, 5, c(50, 50, 50, 50, 50), 100)
PV(0.01, 3, c(50, 100, 150), 100)
PV(0.03, 10, 0, 5000)

}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, March 2021
}
\keyword{discounting}
\keyword{present}
\keyword{value}
