% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RANDBETWEEN.R
\name{RANDBETWEEN}
\alias{RANDBETWEEN}
\title{Get Random Number Between Two Numbers}
\usage{
RANDBETWEEN(low, high)
}
\arguments{
\item{low}{Lowest integer that should be included in the range.}

\item{high}{Highest integer that should be included in the range.}
}
\value{
Returns an integer in the range of numbers.
}
\description{
Given a range of numbers, get a random number in that range. If a non-integer numeric is given, 
will round to the nearest whole number.
}
\examples{
RANDBETWEEN(0, 5)
RANDBETWEEN(-5, 100)

}
\seealso{
\code{\link{sample}}
}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, August 2021
}
\keyword{random}
\keyword{sample}
