% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIGHT.R
\name{RIGHT}
\alias{RIGHT}
\title{Return Specified Number of Characters from Right in String/Vector}
\usage{
RIGHT(x, length)
}
\arguments{
\item{x}{Character vector or string you would like to slice.}

\item{length}{Number of characters from right you would like returned.}
}
\value{
Returns the number of characters from right in a string/character vector.
}
\description{
Returns the number of characters from right in a string/character vector.
}
\examples{
RIGHT("hello world!", 6)
RIGHT(c("hi", "hey", "hello"), 2)

}
\seealso{
\code{\link{substr}}
}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, January 2021
}
\keyword{substr}
