% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUMIF.R
\name{SUMIF}
\alias{SUMIF}
\title{Sums Vector Components Based On Condition}
\usage{
SUMIF(sum_vector, condition_vector, condition)
}
\arguments{
\item{sum_vector}{Vector that you would like to sum.}

\item{condition_vector}{Vector that you would like condition specified for.}

\item{condition}{Value that will satisfy certain components in condition vector.}
}
\value{
Returns summed value in the sum vector based on condition specified for condition vector.
}
\description{
Takes two vectors, one being the vector you want to sum, and the other
being the vector you want create a condition for. Then, specify the condition
that should be used to satisfy the condition vector and it will aggregate the sum vector accordingly.
}
\examples{
SUMIF(iris$Sepal.Length, iris$Species, "virginica")
SUMIF(c(1, 2, 3), c(3, 1, 3), 3)

}
\seealso{
\code{\link{sum}}
}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, September 2020
}
\keyword{conditional}
\keyword{sum}
