% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TODAY.R
\name{TODAY}
\alias{TODAY}
\title{Get Today's Date}
\usage{
TODAY()
}
\value{
Returns today's date as a date object.
}
\description{
Returns today's date.
}
\examples{
TODAY()

}
\seealso{
\code{\link{Sys.Date}}
}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, February 2021
}
\keyword{date}
\keyword{time}
\keyword{today}
