% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRIM.R
\name{TRIM}
\alias{TRIM}
\title{Trim Leading & Trailing Whitespace in String/Vector}
\usage{
TRIM(x)
}
\arguments{
\item{x}{Character vector or string you would like to remove trailing/leading whitespace from.}
}
\value{
Returns the string/character vector with leading/trailing whitespace removed.
}
\description{
Trims leading/trailing whitespace in a string/character vector.
}
\examples{
TRIM("     hello world!    ")
TRIM(c("hi    ", "    hey", "   hello   "))

}
\seealso{
\code{\link{trimws}}
}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, January 2021
}
\keyword{trim}
\keyword{trimws}
