% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UPPER.R
\name{UPPER}
\alias{UPPER}
\title{Convert Text to Uppercase}
\usage{
UPPER(text)
}
\arguments{
\item{text}{String or vector that you would like converted to uppercase text.}
}
\value{
Returns the string or vector in uppercase text.
}
\description{
Given a string or vector, convert to uppercase text.
}
\examples{
UPPER("hello")
UPPER(c("hey", "HI", "hey"))
UPPER(c(1, "hi", "hey", "HI"))

}
\seealso{
\code{\link{toupper}}
}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, January 2021
}
\keyword{toupper}
\keyword{upper}
