% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlookup.R
\name{vlookup}
\alias{vlookup}
\title{Find/Return Values From Matching Rows in Data Frame.}
\usage{
vlookup(lookup_vector, lookup_df, join_col_name, column_index)
}
\arguments{
\item{lookup_vector}{Values that you would like to join on.}

\item{lookup_df}{Data frame that contains values in lookup_vector and other values you would like to connect.}

\item{join_col_name}{Name of the column (as a string) that has values from lookup_vector.}

\item{column_index}{Column in data frame that you would like returned once lookup value is found in lookup_df.}
}
\value{
Dataframe of returning values matching the lookup vector as well as the lookup vector values.
}
\description{
Finds matching rows in specific column of data frame that matches the values specified.
}
\examples{
vlookup(c("Nick", "Jake", "Tyler", "Nate"), data.frame(name = c("Nick", "Jake", "Tyler", "Nate"), number = 1:4), "name", 2)
}
\seealso{
\code{\link{merge}}
}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, September 2020
}
\keyword{join}
\keyword{left}
\keyword{merge}
