% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COUNT_IF.R
\name{COUNT_IF}
\alias{COUNT_IF}
\title{Count Vector Components Based On Condition}
\usage{
COUNT_IF(condition_vector, condition)
}
\arguments{
\item{condition_vector}{Vector that you would like condition specified for.}

\item{condition}{Value that will satisfy certain components in condition vector.}
}
\value{
Returns count in the condition vector based on the condition specified.
}
\description{
Counts the amount of instances a condition is specified in a vector.
}
\examples{
COUNT_IF(iris$Species, "virginica")
COUNT_IF(c(3, 1, 3), 3)

}
\seealso{
\code{\link{sum}}
}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, September 2020
}
\keyword{conditional}
\keyword{count}
